Logik neu + Anbindung an die Oberfläche (vorläufig)

+++ ProjektKlausurgenerator/src/de/arbeitsagentur/ProjektKlausurgenerator/model/AbstractFrage.java
@@ -14,7 +14,7 @@ import de.arbeitsagentur.ProjektKlausurgenerator.enums.Schwierigkeitsgrad;
 public abstract class AbstractFrage {
 	protected String frageText;
 	protected Schwierigkeitsgrad schwierigkeitsgrad;
-	protected int punkte;
+	protected Double punkte;
 	protected String seminar;
 
 	/**
@@ -25,7 +25,7 @@ public abstract class AbstractFrage {
 	 * @param punkte
 	 * @param seminar
 	 */
-	AbstractFrage(String frage, Schwierigkeitsgrad schwierigkeitsgrad, int punkte, String seminar) {
+	AbstractFrage(String frage, Schwierigkeitsgrad schwierigkeitsgrad, Double punkte, String seminar) {
 		frageText = frage;
 		this.schwierigkeitsgrad = schwierigkeitsgrad;
 		this.punkte = punkte;
@@ -36,7 +36,7 @@ public abstract class AbstractFrage {
 	AbstractFrage(String[] rawFrage) {
 		frageText = rawFrage[1];
 		schwierigkeitsgrad = Schwierigkeitsgrad.valueOf(rawFrage[2]);
-		punkte = Integer.valueOf(rawFrage[3]);
+		punkte = Double.valueOf(rawFrage[3]);
 		seminar = rawFrage[4];
 	}
 
@@ -48,7 +48,7 @@ public abstract class AbstractFrage {
 		return schwierigkeitsgrad;
 	}
 
-	public int getPunkte() {
+	public Double getPunkte() {
 		return punkte;
 	}
